requirements.txt
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
MarkupSafe==3.0.2
python-dotenv==1.0.1 